server/requirements.txt
fastapi==0.117.0
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
SQLAlchemy==2.0.38
alembic==1.14.1
psycopg[binary]==3.2.9
PyJWT==2.10.1
argon2-cffi==23.1.0